size_t

+++ 10-binary_tree_depth.c
@@ -8,7 +8,7 @@ size_t binary_tree_depth(const binary_tree_t *tree)
 {
 	if (tree->parent == NULL)
 		return (0);
-	int n = binary_tree_depth(tree->parent);
+	size_t n = binary_tree_depth(tree->parent);
 
 	return (n + 1);
 }
